backend/pyproject.toml
[project]
name = "ybryx-backend"
version = "0.1.0"
description = "Ybryx Capital Robotics Financing Backend - Multi-Agent LangGraph System"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & ASGI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # LangChain & LangGraph
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.0.40",
    "langchain-community>=0.0.20",

    # Database & ORM
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",

    # Supabase & Auth
    "supabase>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Memory & Vector Stores
    "mem0ai>=0.0.5",
    "chromadb>=0.4.20",
    "qdrant-client>=1.7.0",

    # Data Validation & Settings
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # HTTP Client
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",

    # Logging & Monitoring
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",

    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "faker>=22.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
    "ipdb>=0.13.13",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_backend"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
